% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col.R
\name{add_col}
\alias{add_col}
\title{Add column into an existing datafrma}
\usage{
add_col(df1, name, value, where)
}
\arguments{
\item{df1}{A dataframe}

\item{name}{A character arg}

\item{value}{something}

\item{where}{Integer value}
}
\value{

}
\description{
Cool, huh?
}
