% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into.R
\name{insert_into}
\alias{insert_into}
\title{Insert one dataframe into another}
\usage{
insert_into(x, y, where = 1)
}
\arguments{
\item{x}{A dataframe}

\item{y}{A dataframe}

\item{where}{An integer value}
}
\value{
A dataframe
}
\description{
You can specify where to splice in
}
